consolidated all template checking to a single place. There is now only one call to bencode, bdecode, and a btemplatein the entire program.
wrote tests for Downloader and fixed bugs found as a result. This was notable
from several points of view - the tests were much harder to write than
expected, the number of bugs found was much greater than expected, and
the severity of the found bugs was much greater than expected.

Apparently the 'fix everything' design pattern used in the choker and
downloader results in very short and readable code, but when it contains
bugs they're often extremely subtle and have very bizarre counterintuitive
effects.