improved messages
A couple people reported a 'bug' in which resuming hung - it turns out that
scanning a 342 meg file takes a minute or two on a lot of hard drives, and it
was working fine, but they gave up on it before that operation was complete.

Now it says that scanning and allocation may take a while, which will hopefully sate peoples's impatience. Really it should give a progress display on file scanning or allocation - I'll do the progress display on scanning later, although I'm not sure what to do about allocation. Python has this clever trick which will allocate sparse files if you scan to the end and write a single byte out, but I believe on systems which don't support sparse files it manually allocates the whole thing, which may take a while, I'm not sure. In any case, that API doesn't supply an easy way of displaying a progress meter.

I don't know of any place where that trick is documented - I found it after judicious web searching in a regression test which claimed to be for checking Python's sparse file handling.