CVS, meet btlaunchmanycurses.py.
btlaunchmanycurses.py is a many-torrent client which monitors a directory for
.torrent files, and attempts to start up a torrent assuming anything that is
named <something>.torrent is a responsefile.  If it fails (usually because the
responsefile isn't valid), the file is marked as "dead" and is not tried again.
btlaunchmanycurses tracks changes in the directory, and such, will stop
torrents whose corresponding files have been removed, and start torrents for
new files appearing in the directory.

btlaunchmanycurses.py is executed and takes the same arguments as
btlaunchmany.py.

btlaunchmany.py gained the same directory tracking code that exists in
btlaunchmanycurses.py.